+++ COVID19Py/covid19.py
@@ -1,6 +1,7 @@
 from typing import Dict, List
 import requests
 import json
+from info import Info
 
 class COVID19(object):
     default_url = "https://covid-tracker-us.herokuapp.com"
@@ -33,7 +34,7 @@ class COVID19(object):
                     self.url = ""
                     continue
 
-                # TODO: Should have a better health-check, this is way too hacky...
+                # Should have a better health-check, this is way too hacky...
                 if "jhu" in result:
                     # We found a mirror that worked just fine, let's stick with it
                     break
@@ -48,6 +49,8 @@ class COVID19(object):
         if data_source not in self._valid_data_sources:
             raise ValueError("Invalid data source. Expected one of: %s" % self._valid_data_sources)
         self.data_source = data_source
+ 
+        self.obj = Info(self.url, self.data_source)
 
     def _update(self, timelines):
         latest = self.getLatest()
@@ -123,36 +126,17 @@ class COVID19(object):
 
         return data
 
-    def getLocationByCountryCode(self, country_code, timelines=False) -> List[Dict]:
-        """
-        :param country_code: String denoting the ISO 3166-1 alpha-2 code (https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2) of the country
-        :param timelines: Whether timeline information should be returned as well.
-        :return: A list of areas that correspond to the country_code. If the country_code is invalid, it returns an empty list.
-        """
-        data = None
-        if timelines:
-            data = self._request("/v2/locations", {"country_code": country_code, "timelines": str(timelines).lower()})
-        else:
-            data = self._request("/v2/locations", {"country_code": country_code})
-        return data["locations"]
-    
-    def getLocationByCountry(self, country, timelines=False) -> List[Dict]:
-        """
-        :param country: String denoting name of the country
-        :param timelines: Whether timeline information should be returned as well.
-        :return: A list of areas that correspond to the country name. If the country is invalid, it returns an empty list.
-        """
-        data = None
-        if timelines:
-            data = self._request("/v2/locations", {"country": country, "timelines": str(timelines).lower()})
-        else:
-            data = self._request("/v2/locations", {"country": country})
-        return data["locations"]
-
-    def getLocationById(self, country_id: int):
-        """
-        :param country_id: Country Id, an int
-        :return: A dictionary with case information for the specified location.
-        """
-        data = self._request("/v2/locations/" + str(country_id))
-        return data["location"]
+    #this method will accept either int or string input to fetch info (it will replace having to interact with 3 diff methods)
+    #it first checks for integer input, then moves onto trying string input, then if it is string input it
+    # has to decipher whether its a code or country name 
+    def getLocationBy(self, inp, timelines = False):
+        inp = inp
+        timelines = timelines
+        try:
+            int(inp)
+            return self.obj.getLocationById(inp)
+        except:
+            str(inp)
+            if(len(inp) > 2): #codes are only strings of size 2
+                return self.obj.getLocationByCountry(inp, timelines)
+            return self.obj.getLocationByCountryCode(inp, timelines)
